Module sodaq_one::atsamd21g18a::ac::compctrl

Comparator Control n

Structs

ENABLER
Value of the field

HYSTR
Value of the field

R
Value read from the register

SINGLER
Value of the field

SWAPR
Value of the field

W
Value to write to the register

_ENABLEW
Proxy

_FLENW
Proxy

_HYSTW
Proxy

_INTSELW
Proxy

_MUXNEGW
Proxy

_MUXPOSW
Proxy

_OUTW
Proxy

_SINGLEW
Proxy

_SPEEDW
Proxy

_SWAPW
Proxy

Enums

FLENR
Possible values of the field FLEN

FLENW
Values that can be written to the field FLEN

INTSELR
Possible values of the field INTSEL

INTSELW
Values that can be written to the field INTSEL

MUXNEGR
Possible values of the field MUXNEG

MUXNEGW
Values that can be written to the field MUXNEG

MUXPOSR
Possible values of the field MUXPOS

MUXPOSW
Values that can be written to the field MUXPOS

OUTR
Possible values of the field OUT

OUTW
Values that can be written to the field OUT

SPEEDR
Possible values of the field SPEED

SPEEDW
Values that can be written to the field SPEED