Enum sodaq_one::atsamd21g18a::wdt::ewctrl::EWOFFSETR

pub enum EWOFFSETR {
    _8,
    _16,
    _32,
    _64,
    _128,
    _256,
    _512,
    _1K,
    _2K,
    _4K,
    _8K,
    _16K,
    _Reserved(u8),
}

Possible values of the field EWOFFSET

Variants
_8

8 clock cycles

_16

16 clock cycles

_32

32 clock cycles

_64

64 clock cycles

_128

128 clock cycles

_256

256 clock cycles

_512

512 clock cycles

_1K

1024 clock cycles

_2K

2048 clock cycles

_4K

4096 clock cycles

_8K

8192 clock cycles

_16K

16384 clock cycles

_Reserved(u8)

Reserved

Methods

impl EWOFFSETR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_16(&self) -> bool

Checks if the value of the field is _16

pub fn is_32(&self) -> bool

Checks if the value of the field is _32

pub fn is_64(&self) -> bool

Checks if the value of the field is _64

pub fn is_128(&self) -> bool

Checks if the value of the field is _128

pub fn is_256(&self) -> bool

Checks if the value of the field is _256

pub fn is_512(&self) -> bool

Checks if the value of the field is _512

pub fn is_1k(&self) -> bool

Checks if the value of the field is _1K

pub fn is_2k(&self) -> bool

Checks if the value of the field is _2K

pub fn is_4k(&self) -> bool

Checks if the value of the field is _4K

pub fn is_8k(&self) -> bool

Checks if the value of the field is _8K

pub fn is_16k(&self) -> bool

Checks if the value of the field is _16K