Struct sodaq_one::gpio::Pa24

pub struct Pa24<MODE> { /* fields omitted */ }

Represents the IO pin with the matching name.

Methods

impl<MODE> Pa24<MODE>

pub fn into_floating_input(self, port: &mut Port) -> Pa24<Input<Floating>>

Configures the pin to operate as a floating input

pub fn into_pull_down_input(self, port: &mut Port) -> Pa24<Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self, port: &mut Port) -> Pa24<Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_open_drain_output(self, port: &mut Port) -> Pa24<Output<OpenDrain>>

Configures the pin to operate as an open drain output

pub fn into_push_pull_output(self, port: &mut Port) -> Pa24<Output<PushPull>>

Configures the pin to operate as a push-pull output

impl Pa24<Output<OpenDrain>>

pub fn internal_pull_up(&mut self, port: &mut Port, on: bool)

Control state of the internal pull up

impl<MODE> Pa24<Output<MODE>>

pub fn toggle(&mut self)

Toggle the logic level of the pin; if it is currently high, set it low and vice versa.

impl<MODE> Pa24<MODE>

pub fn into_function_a(self, port: &mut Port) -> Pa24<PfA>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_b(self, port: &mut Port) -> Pa24<PfB>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_c(self, port: &mut Port) -> Pa24<PfC>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_d(self, port: &mut Port) -> Pa24<PfD>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_e(self, port: &mut Port) -> Pa24<PfE>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_f(self, port: &mut Port) -> Pa24<PfF>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_g(self, port: &mut Port) -> Pa24<PfG>

Configures the pin to operate with a peripheral

impl<MODE> Pa24<MODE>

pub fn into_function_h(self, port: &mut Port) -> Pa24<PfH>

Configures the pin to operate with a peripheral